refactor: example 패키지를 job 하위로 이동
