+++ src/main/java/egovframework/bat/job/example/config/MybatisToMybatisJobConfig.java
@@ -1,4 +1,4 @@
-package egovframework.bat.example.config;
+package egovframework.bat.job.example.config;
 
 import org.egovframe.rte.bat.core.item.database.EgovMyBatisBatchItemWriter;
 import org.egovframe.rte.bat.core.item.database.EgovMyBatisPagingItemReader;
@@ -16,8 +16,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
-import egovframework.bat.example.domain.CustomerCredit;
-import egovframework.bat.example.processor.CustomerCreditIncreaseProcessor;
+import egovframework.bat.job.example.domain.CustomerCredit;
+import egovframework.bat.job.example.processor.CustomerCreditIncreaseProcessor;
 
 import org.apache.ibatis.session.SqlSessionFactory;
