+++ src/main/java/egovframework/bat/job/example/processor/CustomerCreditIncreaseProcessor.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package egovframework.bat.example.processor;
+package egovframework.bat.job.example.processor;
 
 import java.math.BigDecimal;
 
 import org.springframework.batch.item.ItemProcessor;
-import egovframework.bat.example.domain.CustomerCredit;
+import egovframework.bat.job.example.domain.CustomerCredit;
 
 /**
  * @author 배치실행개발팀
