+++ src/main/resources/egovframework/batch/context-batch-mapper.xml
@@ -28,7 +28,7 @@
                 <value>classpath:/egovframework/batch/mapper/insa/insa_remote1_to_stg.xml</value>
                 <value>classpath:/egovframework/batch/mapper/job/erp/erp_stg_to_local.xml</value>
                 <value>classpath:/egovframework/batch/mapper/job/erp/erp_rest_to_stg.xml</value>
-                <value>classpath:/egovframework/batch/mapper/example/Egov_Example_SQL.xml</value>
+                <value>classpath:/egovframework/batch/mapper/job/example/Egov_Example_SQL.xml</value>
                 <!-- 배치 관리 매퍼는 스테이징 환경에서만 사용됩니다. -->
                 <value>classpath:/egovframework/batch/mapper/batch/BatchManagementMapper.xml</value>
             </list>
@@ -52,7 +52,7 @@
         <property name="configLocation" value="classpath:/egovframework/batch/mapper/config/mapper-config.xml" />
         <property name="mapperLocations">
             <list>
-                <value>classpath:/egovframework/batch/mapper/example/Egov_Example_SQL.xml</value>
+                <value>classpath:/egovframework/batch/mapper/job/example/Egov_Example_SQL.xml</value>
             </list>
         </property>
     </bean>
